# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DotAAS Part 2 | HTTP/REST | Entire Interface Collection",
    "description" : "The entire interface collection as part of Details of the Asset Administration Shell Part 2",
    "termsOfService" : "https://github.com/admin-shell-io/aas-specs",
    "contact" : {
      "name" : "Michael Hoffmeister, Torben Miny, Andreas Orzelski, Manuel Sauer, Constantin Ziesche"
    },
    "version" : "V1.0RC01"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/aas" : {
      "get" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Returns the Asset Administration Shell",
        "operationId" : "GetAssetAdministrationShell",
        "parameters" : [ {
          "name" : "content",
          "in" : "query",
          "description" : "Determines the request or response kind of the resource",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "normal", "trimmed", "value", "reference", "path" ],
            "default" : "normal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested Asset Administration Shell",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssetAdministrationShell"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetAssetAdministrationShell/1/0/RC02" ]
      },
      "put" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Updates the Asset Administration Shell",
        "operationId" : "PutAssetAdministrationShell",
        "parameters" : [ {
          "name" : "content",
          "in" : "query",
          "description" : "Determines the request or response kind of the resource",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "normal", "trimmed", "value", "reference", "path" ],
            "default" : "normal"
          }
        } ],
        "requestBody" : {
          "description" : "Asset Administration Shell object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssetAdministrationShell"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Asset Administration Shell updated successfully"
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PutAssetAdministrationShell/1/0/RC02" ]
      }
    },
    "/aas/asset-information" : {
      "get" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Returns the Asset Information",
        "operationId" : "GetAssetInformation",
        "responses" : {
          "200" : {
            "description" : "Requested Asset Information",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssetInformation"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetAssetInformation/1/0/RC02" ]
      },
      "put" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Updates the Asset Information",
        "operationId" : "PutAssetInformation",
        "requestBody" : {
          "description" : "Asset Information object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssetInformation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Asset Information updated successfully"
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PutAssetInformation/1/0/RC02" ]
      }
    },
    "/aas/submodels" : {
      "get" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Returns all submodel references",
        "operationId" : "GetAllSubmodelReferences",
        "responses" : {
          "200" : {
            "description" : "Requested submodel references",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Reference"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetAllSubmodelReferences/1/0/RC02" ]
      },
      "post" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Creates a submodel reference at the Asset Administration Shell",
        "operationId" : "PostSubmodelReference",
        "requestBody" : {
          "description" : "Reference to the Submodel",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Submodel reference created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Reference"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PostSubmodelReference/1/0/RC02" ]
      }
    },
    "/aas/submodels/{submodelIdentifier}" : {
      "delete" : {
        "tags" : [ "Asset Administration Shell Interface" ],
        "summary" : "Deletes the submodel reference from the Asset Administration Shell",
        "operationId" : "DeleteSubmodelReferenceById",
        "parameters" : [ {
          "name" : "submodelIdentifier",
          "in" : "path",
          "description" : "The Submodel’s unique id (BASE64-URL-encoded)",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Submodel reference deleted successfully"
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/DeleteSubmodelReference/1/0/RC02" ]
      }
    },
...
    "/serialization" : {
      "get" : {
        "tags" : [ "Asset Administration Shell Serialization Interface" ],
        "summary" : "Returns an appropriate serialization based on the specified format (see SerializationFormat)",
        "operationId" : "GenerateSerializationByIds",
        "parameters" : [ {
          "name" : "aasIds",
          "in" : "query",
          "description" : "The Asset Administration Shells' unique ids (BASE64-URL-encoded)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        }, {
          "name" : "submodelIds",
          "in" : "query",
          "description" : "The Submodels' unique ids (BASE64-URL-encoded)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        }, {
          "name" : "includeConceptDescriptions",
          "in" : "query",
          "description" : "Include Concept Descriptions?",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested serialization based on SerializationFormat",
            "content" : {
              "application/asset-administration-shell-package+xml" : {
                "schema" : {
                  "type" : "string",
                  "description" : "AASX package",
                  "format" : "binary",
                  "x-content-type" : "application/asset-administration-shell-package+xml"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssetAdministrationShellEnvironment"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssetAdministrationShellEnvironment"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GenerateSerializationByIds/1/0/RC02" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "IdentifierKeyValuePair" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "required" : [ "key", "subjectId", "value" ],
          "properties" : {
            "key" : {
              "type" : "string"
            },
            "subjectId" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "value" : {
              "type" : "string"
            }
          }
        } ]
      },
      "HasSemantics" : {
        "type" : "object",
        "properties" : {
          "semanticId" : {
            "$ref" : "#/components/schemas/Reference"
          }
        }
      },
      "Reference" : {
        "type" : "object",
        "oneOf" : [ {
          "$ref" : "#/components/schemas/GlobalReference"
        }, {
          "$ref" : "#/components/schemas/ModelReference"
        } ]
      },
      "GlobalReference" : {
        "required" : [ "value" ],
        "properties" : {
          "value" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        }
      },
      "Identifier" : {
        "type" : "string"
      },
      "ModelReference" : {
        "required" : [ "keys" ],
        "properties" : {
          "referredSemanticId" : {
            "$ref" : "#/components/schemas/GlobalReference"
          },
          "keys" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Key"
            }
          }
        }
      },
      "Key" : {
        "required" : [ "type", "value" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/KeyElements"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
      "KeyElements" : {
        "type" : "string",
        "enum" : [ "AssetAdministrationShell", "AccessPermissionRule", "ConceptDescription", "Submodel", "AnnotatedRelationshipElement", "BasicEvent", "Blob", "Capability", "DataElement", "File", "Entity", "Event", "MultiLanguageProperty", "Operation", "Property", "Range", "ReferenceElement", "RelationshipElement", "SubmodelElement", "SubmodelElementList", "SubmodelElementStruct", "View", "FragmentReference" ]
      },
      "AssetAdministrationShell" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "required" : [ "assetInformation" ],
          "properties" : {
            "assetInformation" : {
              "$ref" : "#/components/schemas/AssetInformation"
            },
            "derivedFrom" : {
              "$ref" : "#/components/schemas/Reference"
            },
            "security" : {
              "$ref" : "#/components/schemas/Security"
            },
            "submodels" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            },
            "views" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/View"
              }
            }
          }
        } ]
      },
      "Identifiable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Referable"
        }, {
          "required" : [ "identification" ],
          "properties" : {
            "administration" : {
              "$ref" : "#/components/schemas/AdministrativeInformation"
            },
            "identification" : {
              "$ref" : "#/components/schemas/Identifier"
            }
          }
        } ]
      },
      "Referable" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/HasExtensions"
        }, {
          "required" : [ "idShort", "modelType" ],
          "properties" : {
            "category" : {
              "type" : "string"
            },
            "description" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "displayName" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/LangString"
              }
            },
            "idShort" : {
              "type" : "string"
            },
            "modelType" : {
              "$ref" : "#/components/schemas/ModelType"
            }
          }
        } ]
      },
      "HasExtensions" : {
        "type" : "object",
        "properties" : {
          "extensions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Extension"
            }
          }
        }
      },
...
      "LangString" : {
        "required" : [ "language", "text" ],
        "type" : "object",
        "properties" : {
          "language" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        },
        "example" : {
          "language" : "language",
          "text" : "text"
        }
      },
      "ModelType" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "$ref" : "#/components/schemas/ModelTypes"
          }
        }
      },
...
      "AdministrativeInformation" : {
        "type" : "object",
        "properties" : {
          "revision" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        },
        "example" : {
          "version" : "version",
          "revision" : "revision"
        }
      },
      "HasDataSpecification" : {
        "type" : "object",
        "properties" : {
          "embeddedDataSpecifications" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmbeddedDataSpecification"
            }
          }
        }
      },
      "EmbeddedDataSpecification" : {
        "required" : [ "dataSpecification", "dataSpecificationContent" ],
        "type" : "object",
        "properties" : {
          "dataSpecification" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "dataSpecificationContent" : {
            "$ref" : "#/components/schemas/DataSpecificationContent"
          }
        }
      },
      "DataSpecificationContent" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/DataSpecificationIEC61360Content"
        }, {
          "$ref" : "#/components/schemas/DataSpecificationPhysicalUnitContent"
        } ]
      },
...
      "AssetInformation" : {
        "required" : [ "assetKind" ],
        "properties" : {
          "assetKind" : {
            "$ref" : "#/components/schemas/AssetKind"
          },
          "billOfMaterial" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "globalAssetId" : {
            "$ref" : "#/components/schemas/Reference"
          },
          "specificAssetIds" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IdentifierKeyValuePair"
            }
          },
          "thumbnail" : {
            "$ref" : "#/components/schemas/File"
          }
        },
        "example" : {
          "thumbnail" : "",
          "assetKind" : "Type",
          "specificAssetIds" : [ "", "" ],
          "billOfMaterial" : [ "", "" ],
          "globalAssetId" : null
        }
      },
      "AssetKind" : {
        "type" : "string",
        "enum" : [ "Type", "Instance" ]
      },
      "File" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SubmodelElementAttributes"
        }, {
          "required" : [ "mimeType" ],
          "properties" : {
            "mimeType" : {
              "type" : "string"
            },
            "value" : {
              "type" : "string"
            }
          }
        } ]
      },
      "SubmodelElementAttributes" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Referable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "$ref" : "#/components/schemas/Qualifiable"
        }, {
          "properties" : {
            "kind" : {
              "$ref" : "#/components/schemas/ModelingKind"
            }
          }
        } ]
      },
      "Qualifiable" : {
        "type" : "object",
        "properties" : {
          "qualifiers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Constraint"
            }
          }
        }
      },
      "Constraint" : {
        "required" : [ "modelType" ],
        "properties" : {
          "modelType" : {
            "$ref" : "#/components/schemas/ModelType"
          }
        },
        "oneOf" : [ {
          "$ref" : "#/components/schemas/Qualifier"
        }, {
          "$ref" : "#/components/schemas/Formula"
        } ]
      },
...
      "ModelingKind" : {
        "type" : "string",
        "enum" : [ "Template", "Instance" ]
      },
      "Security" : {
        "required" : [ "accessControlPolicyPoints" ],
        "type" : "object",
        "properties" : {
          "accessControlPolicyPoints" : {
            "$ref" : "#/components/schemas/AccessControlPolicyPoints"
          },
          "certificate" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Certificate"
            }
          },
          "requiredCertificateExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          }
        }
      },
      "AccessControlPolicyPoints" : {
        "required" : [ "policyAdministrationPoint", "policyDecisionPoint", "policyEnforcementPoint" ],
        "type" : "object",
        "properties" : {
          "policyAdministrationPoint" : {
            "$ref" : "#/components/schemas/PolicyAdministrationPoint"
          },
          "policyDecisionPoint" : {
            "$ref" : "#/components/schemas/PolicyDecisionPoint"
          },
          "policyEnforcementPoint" : {
            "$ref" : "#/components/schemas/PolicyEnforcementPoint"
          },
          "policyInformationPoints" : {
            "$ref" : "#/components/schemas/PolicyInformationPoints"
          }
        }
      },
      "PolicyAdministrationPoint" : {
        "required" : [ "externalAccessControl" ],
        "type" : "object",
        "properties" : {
          "externalAccessControl" : {
            "type" : "boolean"
          },
          "localAccessControl" : {
            "$ref" : "#/components/schemas/AccessControl"
          }
        }
      },
...
      "PolicyDecisionPoint" : {
        "required" : [ "externalPolicyDecisionPoints" ],
        "type" : "object",
        "properties" : {
          "externalPolicyDecisionPoints" : {
            "type" : "boolean"
          }
        }
      },
      "PolicyEnforcementPoint" : {
        "required" : [ "externalPolicyEnforcementPoint" ],
        "type" : "object",
        "properties" : {
          "externalPolicyEnforcementPoint" : {
            "type" : "boolean"
          }
        }
      },
      "PolicyInformationPoints" : {
        "required" : [ "externalInformationPoint" ],
        "type" : "object",
        "properties" : {
          "externalInformationPoint" : {
            "type" : "boolean"
          },
          "internalInformationPoint" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          }
        }
      },
      "Certificate" : {
        "type" : "object",
        "properties" : {
          "blobCertificate" : {
            "$ref" : "#/components/schemas/Blob"
          },
          "containedExtension" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Reference"
            }
          },
          "lastCertificate" : {
            "type" : "boolean"
          }
        }
      },
      "View" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Referable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "properties" : {
            "containedElements" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            }
          }
        } ]
      },
      "Submodel" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "$ref" : "#/components/schemas/Qualifiable"
        }, {
          "$ref" : "#/components/schemas/HasSemantics"
        }, {
          "properties" : {
            "kind" : {
              "$ref" : "#/components/schemas/ModelingKind"
            },
            "submodelElements" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/SubmodelElement"
              }
            }
          }
        } ]
      },
      "SubmodelElement" : {
        "oneOf" : [ {
          "$ref" : "#/components/schemas/Blob"
        }, {
          "$ref" : "#/components/schemas/File"
        }, {
          "$ref" : "#/components/schemas/Capability"
        }, {
          "$ref" : "#/components/schemas/Entity"
        }, {
          "$ref" : "#/components/schemas/Event"
        }, {
          "$ref" : "#/components/schemas/BasicEvent"
        }, {
          "$ref" : "#/components/schemas/MultiLanguageProperty"
        }, {
          "$ref" : "#/components/schemas/Operation"
        }, {
          "$ref" : "#/components/schemas/Property"
        }, {
          "$ref" : "#/components/schemas/Range"
        }, {
          "$ref" : "#/components/schemas/ReferenceElement"
        }, {
          "$ref" : "#/components/schemas/AnnotatedRelationshipElement"
        }, {
          "$ref" : "#/components/schemas/RelationshipElement"
        }, {
          "$ref" : "#/components/schemas/SubmodelElementList"
        }, {
          "$ref" : "#/components/schemas/SubmodelElementStruct"
        } ]
      },
      "Blob" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SubmodelElementAttributes"
        }, {
          "required" : [ "mimeType" ],
          "properties" : {
            "mimeType" : {
              "type" : "string"
            },
            "value" : {
              "type" : "string"
            }
          }
        } ]
      },
...
      "Event" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SubmodelElementAttributes"
        } ]
      },
...
      "ConceptDescription" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Identifiable"
        }, {
          "$ref" : "#/components/schemas/HasDataSpecification"
        }, {
          "properties" : {
            "isCaseOf" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Reference"
              }
            }
          }
        } ]
      },
...
      "AssetAdministrationShellEnvironment" : {
        "type" : "object",
        "properties" : {
          "assetAdministrationShells" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AssetAdministrationShell"
            }
          },
          "conceptDescriptions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConceptDescription"
            }
          },
          "submodels" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Submodel"
            }
          }
        }
      },
...
    }
  }
}